Remove `HiveTableRedirectionsProvider` interface
Due to the ability of the `hive` connector to perform table
redirections towards both `iceberg` and `delta-lake` tables,
there is no need anymore for providing an extension point
to be able to customize the way in which table redirections
are made.
